@Noiredd 2017-06-14
Lint and runtests on my machine have all succeeded. Travis fails some numeric check with `FAIL: test_backward_start_end (test_net.TestNet)`, details of the error are a little [cryptic](https://travis-ci.org/BVLC/caffe/jobs/242710271#L8685). However, [previous builds](https://travis-ci.org/BVLC/caffe/builds/242660340) have already failed at the similar stage (different test jobs though), therefore I suspect that the problem is somewhere else.  
Comments?