@Noiredd 2017-06-19
Question regarding failed tests: I just noticed that when I `make runtest` even on my machine, sometimes I will get 2101 of 2101 test passed, and sometimes one or two will fail; notably these:  
`[  FAILED  ] AccuracyLayerTest/0.TestForwardIgnoreLabel, where TypeParam = float`  
`[  FAILED  ] AccuracyLayerTest/0.TestForwardCPU, where TypeParam = float`  
How can this be? For some seeds (20620, 58463, 62682) everything looks fine, while for example 74046 fails. Then I update my fork, PR runs tests and they fail. Is there something I don't understand about automated tests?