@Noiredd 2017-06-21
Is there any way I can get to the details of the Travis [build error](https://travis-ci.org/BVLC/caffe/jobs/245295580#L8683)? My `make runtest` passed without problems, but there I got a "100% mismatch", which seems confusing because the visible parts of `x` and `y` arrays look identical.