@Noiredd 2017-07-03
I guess I force-pushed too much and broke Travis. Going to close this and open a new PR.